backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
supabase==2.10.0
python-dotenv==1.0.0
python-multipart==0.0.9
pydantic==2.12.5

